requirements.txt
fastapi
uvicorn
google-cloud-storage
numpy
scikit-learn
joblib